Move local/basic strategy out of routes
