+++ server/routes/local-auth.js
@@ -1,11 +1,7 @@
 const passport = require('passport');
-const PassportLocalStrategy = require('passport-local').Strategy;
-const BasicStrategy = require('passport-http').BasicStrategy;
 const express = require('express');
 const checkWhitelist = require('../lib/check-whitelist');
 const sendError = require('../lib/send-error');
-const appLog = require('../lib/app-log');
-const passhash = require('../lib/passhash.js');
 
 async function handleSignup(req, res, next) {
   const { models } = req;
@@ -65,78 +61,6 @@ function makeLocalAuth(config) {
   const router = express.Router();
 
   if (!config.get('disableUserpassAuth')) {
-    // Register local auth strategy
-    // TODO: Should all authentication strategies be disabled by default, requiring SQLPad implementer to pick one?
-    // This might be more secure as it'd prevent SQLPad's from launching with local auth enabled and open admin registration by default
-    appLog.info('Enabling local authentication strategy.');
-    passport.use(
-      new PassportLocalStrategy(
-        {
-          passReqToCallback: true,
-          usernameField: 'email'
-        },
-        async function passportLocalStrategyHandler(
-          req,
-          email,
-          password,
-          done
-        ) {
-          try {
-            const { models } = req;
-            const user = await models.users.findOneByEmail(email);
-            if (!user) {
-              return done(null, false, { message: 'wrong email or password' });
-            }
-            const isMatch = await passhash.comparePassword(
-              password,
-              user.passhash
-            );
-            if (isMatch) {
-              return done(null, {
-                id: user._id,
-                _id: user._id,
-                role: user.role,
-                email: user.email
-              });
-            }
-            return done(null, false, { message: 'wrong email or password' });
-          } catch (error) {
-            done(error);
-          }
-        }
-      )
-    );
-
-    // TODO - basic auth should perhaps be something opted into as a separate config
-    appLog.info('Enabling basic authentication strategy.');
-    passport.use(
-      new BasicStrategy(
-        {
-          passReqToCallback: true
-        },
-        async function(req, username, password, callback) {
-          try {
-            const { models } = req;
-            const user = await models.users.findOneByEmail(username);
-            if (!user) {
-              return callback(null, false);
-            }
-            const isMatch = await passhash.comparePassword(
-              password,
-              user.passhash
-            );
-            if (!isMatch) {
-              return callback(null, false);
-            }
-            return callback(null, user);
-          } catch (error) {
-            callback(error);
-          }
-        }
-      )
-    );
-
-    // Add routes for local auth signup & signin
     router.post(
       '/api/signup',
       handleSignup,
